refactor: rename vars and file
